Renames samples from kvI to match that of genomic dataset
As kvI (virus) was mapped separately all samples had a different naming scheme from that of the genomic mappings. As some samples of the genomic mappings had to be remapped, these names were alike those of kvI (virus database), thus a need for updating only relevant samples was to be met.